% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_valuation.R
\name{create_home_price_to_income_multiple}
\alias{create_home_price_to_income_multiple}
\title{Home price to income multiple}
\usage{
create_home_price_to_income_multiple()
}
\description{
The `create_home_price_to_income_multiple` function
creates a quotient representing 2019 household income divided
by the Zillow median home price.
}
