% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{fetch_bachelor_degrees}
\alias{fetch_bachelor_degrees}
\title{Fetch percent of population having a bacherlor's degree or higher}
\usage{
fetch_bachelor_degrees()
}
\description{
The `fetch_bachelor_degrees` function retrieves the percentage of the population
having four or more years of college.
}
\references{
\insertRef{USDAeduc2021}{retirementLoc}
}
