% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{fetch_broadband_access}
\alias{fetch_broadband_access}
\title{Fetch percent residents having broadband access}
\usage{
fetch_broadband_access()
}
\description{
The `fetch_broadband_access` function retrieves the percentage of residents
having access to broadband cable.
}
\details{
Broadband coverage is from the US Census Bureau and represents the percentage
 of households having access to broadband in 2017.
}
