% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_county_fips_codes}
\alias{fetch_county_fips_codes}
\title{Fetch county fips codes}
\usage{
fetch_county_fips_codes()
}
\description{
The `fetch_county_fips_codes` function retrieves FIPS codes for 3142 US counties.
}
