% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{fetch_noaa_avg_temp}
\alias{fetch_noaa_avg_temp}
\title{Fetch average annual temparature data}
\usage{
fetch_noaa_avg_temp()
}
\description{
The `fetch_noaa_avg_temp` function retrieves the life expectancy by US county
}
\details{
Average annual temperature was downloaded from \href{https://www.ncdc.noaa.gov/cag/county/mapping}{NOAA National Center for Environmental Information}.
}
