% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{fetch_partisan_lean}
\alias{fetch_partisan_lean}
\title{Fetch Partisan Lean}
\usage{
fetch_partisan_lean()
}
\description{
The `fetch_partisan_lean` function retrieves the percentage of votes by county
cast in favor of the Democratic candidate for president in 2020.
}
\details{
The partisan lean was computed from the 2020 US presidential election results
furnished by the \href{https://electionlab.mit.edu/data}{Election Lab at MIT}.
}
\references{
\insertRef{MIT2017}{retirementLoc}
}
