% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_state_fips_codes}
\alias{fetch_state_fips_codes}
\title{Fetch state fips codes}
\usage{
fetch_state_fips_codes()
}
\description{
The `fetch_state_fips_codes` function retrieves FIPS codes for all 50 US states.
}
