% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{fetch_uscb_population}
\alias{fetch_uscb_population}
\title{Fetch USCB population data}
\usage{
fetch_uscb_population()
}
\description{
The `fetch_uscb_population` function retrieves the total population by county
in the United States for 2020 as tablulated by the United States Census Bureau.
}
\details{
The population data is from the
\href{https://www2.census.gov/programs-surveys/popest/datasets/2010-2020/counties/totals/co-est2020.csv}{U.S. Census Bureau}.
}
\references{
\insertRef{USCB2021}{retirementData}
}
