% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_valuation.R
\name{fetch_usda_household_inc}
\alias{fetch_usda_household_inc}
\title{Fetch USDA Household Income by County}
\usage{
fetch_usda_household_inc()
}
\description{
Fetch USDA Household Income by County
}
\details{
Data may be downloaded from USDA
\href{https://www.ers.usda.gov/data-products/county-level-data-sets/download-data/page}{webpage}.
}
