% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_valuation.R
\name{fetch_zillow_median_price}
\alias{fetch_zillow_median_price}
\title{Fetch Zillow Median Housing Price}
\usage{
fetch_zillow_median_price()
}
\description{
The `fetch_zillow_median_price` function retrieves the typical value for homes
in dollars by county.
}
\details{
Background can be found on Zillow's data \href{https://www.zillow.com/research/data/}{tab}.
According to the Zillow website, it is "a smoothed, seasonally adjusted measure of the typical
home value and market changes across a given region and housing type. It reflects the
typical value for homes in the 35th to 65th percentile range. \insertCite{Zillow2021}{retirementData}
}
\references{
\insertAllCited{}
}
