% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospitalLoc.R
\docType{data}
\name{hospitalLoc}
\alias{hospitalLoc}
\title{US Hospital Locations}
\format{
A data frame with 5746 rows and 8 variables:
\describe{
  \item{lat}{latitude of hospital location}
  \item{lon}{longitude of hospital location}
  \item{name}{hospital name}
  \item{type}{the type of hospital}
  \item{ownership}{the owner of the hospital}
  \item{emer_room}{whether the hospital provides emergency services}
  \item{stars}{the number of stars 1 - 5 that Medicare assigned to the hospital}
  \item{verify}{a not-so-subtle prompt to users to verify the information}
}
}
\usage{
data("hospitalLoc")
}
\description{
A dataset containing the location and rating of US hospitals, VA
medical centers, and military hospitals.
}
\details{
The data was downloaded from
\href{https://data.cms.gov/provider-data/dataset/xubh-q36u}{The Centers for Medicare & Medicaid Services}.
}
\examples{
#load
data("hospitalLoc")
}
\keyword{datasets}
