% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic.R
\name{fetch_broadband_access}
\alias{fetch_broadband_access}
\title{Fetch percent residents having broadband access}
\usage{
fetch_broadband_access()
}
\description{
The `fetch_broadband_access` function retrieves the percentage of residents
having access to broadband cable.
}
\details{
Broadband coverage is from the US Census Bureau and represents the percentage
 of households having access to broadband in 2017. The USCB found that "nationally,
 78 percent of households subscribe to the internet, but households in both rural
 and lower-income counties trail the national average by 13 points."
  \insertCite{USCBbroadband2021}{retirementData} Counties with high broadband
  subscription rates are concentrated around urban areas and along both coasts.
  \insertCite{USCBbroadband2021}{retirementData} Whereas counties with lower subscription
  rates are more prominent in the South and the Mississippi River basin.
  \insertNoCite{USCBacs2018}{retirementData}
}
\references{
\insertAllCited{}
}
