% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housing_valuation.R
\name{fetch_usda_household_inc}
\alias{fetch_usda_household_inc}
\title{Fetch USDA Household Income by County}
\usage{
fetch_usda_household_inc()
}
\description{
The `fetch_usda_household_inc` function retrieves the median household income.
}
\references{
\insertRef{USDAhhinc2021}{retirementData}
}
